requirements.txt
fastapi
google-api-python-client
google-auth
google-generativeai
schedule
uvicorn
sqlite3